Feat : 그룹 생성 컴포넌트 분리
