+++ src/components/icons/BuildingIcon.tsx
@@ -1,0 +1,10 @@
+export const BuildingIcon = () => (
+  <svg width="18" height="18" viewBox="0 0 24 24" fill="none">
+    <path
+      d="M3 21h18M6 21V5a2 2 0 0 1 2-2h8v18M10 7h2M14 7h2M10 11h2M14 11h2M10 15h2M14 15h2"
+      stroke="currentColor"
+      strokeWidth="2"
+      strokeLinecap="round"
+    />
+  </svg>
+);
