+++ src/components/icons/HomeIcon.tsx
@@ -1,0 +1,17 @@
+export const HomeIcon = () => (
+  <svg width="18" height="18" viewBox="0 0 24 24" fill="none">
+    <path
+      d="M3 11l9-7 9 7"
+      stroke="currentColor"
+      strokeWidth="2"
+      strokeLinecap="round"
+    />
+    <path
+      d="M5 10v10h14V10"
+      stroke="currentColor"
+      strokeWidth="2"
+      strokeLinecap="round"
+    />
+    <path d="M9 20v-6h6v6" stroke="currentColor" strokeWidth="2" />
+  </svg>
+);
