+++ src/components/icons/SchoolIcon.tsx
@@ -1,0 +1,7 @@
+export const SchoolIcon = () => (
+  <svg width="18" height="18" viewBox="0 0 24 24" fill="none">
+    <path d="M3 10l9-5 9 5-9 5-9-5z" stroke="currentColor" strokeWidth="2" />
+    <path d="M21 10v6" stroke="currentColor" strokeWidth="2" strokeLinecap="round" />
+    <path d="M3 10v6l9 5 9-5" stroke="currentColor" strokeWidth="2" />
+  </svg>
+);
