+++ src/components/icons/index.ts
@@ -1,0 +1,5 @@
+export * from './PlusIcon';
+export * from './UsersIcon';
+export * from './SchoolIcon';
+export * from './BuildingIcon';
+export * from './HomeIcon';
